Fix issue when venv did not exist and we imported appdirs too early Give anidblink a thread name Rename references to deprecated currentThread to the new current_thread Keeping with the python3 convention of snake_case over camelCase, renamed all references to origThreadName to original_thread_name
