+++ sickchill/views/index.py
@@ -43,7 +43,6 @@ class BaseHandler(RequestHandler):
         super().initialize()
         self.startTime = time.time()
 
-
     # def set_default_headers(self):
     #     self.set_header('Cache-Control', 'no-store, no-cache, must-revalidate, max-age=0')
